requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
--trusted-host pypi.python.org
--trusted-host pypi.org
--trusted-host files.pythonhosted.org

asgiref==3.4.1
    # via django
backports.entry-points-selectable==1.1.1
    # via virtualenv
bump2version==1.0.1
    # via -r requirements.in
cfgv==3.3.1
    # via pre-commit
click==8.0.3
    # via pip-tools
distlib==0.3.3
    # via virtualenv
django==3.2.10
    # via
    #   -r requirements.in
    #   django-classy-tags
    #   django-cms
    #   django-formtools
    #   django-sekizai
    #   django-treebeard
django-classy-tags==2.0.0
    # via
    #   django-cms
    #   django-sekizai
django-cms==3.9.0
    # via
    #   -r requirements.in
    #   djangocms-attributes-field
django-formtools==2.3
    # via django-cms
django-sekizai==2.0.0
    # via django-cms
django-treebeard==4.5.1
    # via django-cms
djangocms-admin-style==2.0.2
    # via django-cms
djangocms-attributes-field==2.0.0
    # via -r requirements.in
filelock==3.4.0
    # via virtualenv
identify==2.4.0
    # via pre-commit
nodeenv==1.6.0
    # via pre-commit
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements.in
platformdirs==2.4.0
    # via virtualenv
pre-commit==2.16.0
    # via -r requirements.in
pytz==2021.3
    # via django
pyyaml==6.0
    # via pre-commit
six==1.16.0
    # via virtualenv
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via pre-commit
tomli==1.2.2
    # via pep517
virtualenv==20.10.0
    # via pre-commit
wheel==0.37.0
    # via
    #   -r requirements.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
